added fix for redir error
