+++ src/parser/handlers/handle_redir_in.c
@@ -12,14 +12,9 @@ int	handle_redir_in(t_token **cur_token, t_cmd **cur_cmd,
 	t_token	*token;
 
 	token = *cur_token;
-	if (!token->next || token->next->type != TOKEN_WORD)
-	{
-		put_error("syntax error near unexpected token `%s'",
-			redir_error(token));
-		minishell->exit_code = E_CRITICAL;
+	if (check_redir_syntax(token, minishell) != EXIT_SUCCESS)
 		return (EXIT_FAILURE);
-	}
-	if (check_ambiguous_redirect(token->next, minishell) == EXIT_FAILURE)
+	if (check_ambiguous_redir(token->next, minishell) == EXIT_FAILURE)
 		return (EXIT_FAILURE);
 	ft_free(1, &(*cur_cmd)->infile);
 	(*cur_cmd)->infile = ft_strdup(token->next->value);
