+++ src/parser/handlers/handle_word.c
@@ -1,5 +1,36 @@
 #include "minishell.h"
 /**
+ * Handles unquoted word tokens by splitting them into individual arguments.
+ * - Creates temporary tokens from split values
+ * - Adds each split part as a separate command argument
+ * - Ensures proper memory cleanup on allocation failure
+ * 
+ * @param cur_cmd Current command being built
+ * @param split_tokens Temporary tokens created from splitting
+ * @param minishell Shell context for error handling
+ */
+static void	handle_unquoted_word(t_cmd *cur_cmd, t_token *split_tokens,
+				t_minishell *minishell)
+{
+	t_token	*tmp;
+	char	*arg_copy;
+
+	tmp = split_tokens;
+	while (tmp)
+	{
+		arg_copy = ft_strdup(tmp->value);
+		if (!arg_copy)
+		{
+			free_token_lst(&split_tokens);
+			exit_minishell(EXIT_FAILURE, minishell, NULL);
+		}
+		add_arg_to_cmd(cur_cmd, arg_copy);
+		tmp = tmp->next;
+	}
+	free_token_lst(&split_tokens);
+}
+
+/**
  * Handles a word token by copying its value and adding it to the current
  * command's arguments.
  * - Duplicates the word and adds it to the args list of the current command.
@@ -12,10 +43,16 @@ int	handle_word(t_token **cur_token, t_cmd **cur_cmd,
 	char	*arg_copy;
 
 	token = *cur_token;
-	arg_copy = ft_strdup(token->value);
-	if (!arg_copy)
-		exit_minishell(EXIT_FAILURE, minishell, NULL);
-	add_arg_to_cmd(*cur_cmd, arg_copy);
+	if (token->was_quoted)
+	{	
+		arg_copy = ft_strdup(token->value);
+		if (!arg_copy)
+			exit_minishell(EXIT_FAILURE, minishell, NULL);
+		add_arg_to_cmd(*cur_cmd, arg_copy);
+	}
+	else
+	handle_unquoted_word(*cur_cmd,
+		split_unquoted(token, token->value), minishell);
 	*cur_token = token->next;
 	return (EXIT_SUCCESS);
 }
