+++ src/tokenizer/token_creation.c
@@ -17,8 +17,6 @@ static t_token	*create_token_with_values(char *original, char *expanded,
 		return (free(original), free(expanded), NULL);
 	token->original_value = original;
 	token->was_quoted = was_quoted;
-	if (!was_quoted && ft_strchr(expanded, ' '))
-		return (split_unquoted(token, expanded));
 	return (token);
 }
 
@@ -81,5 +79,6 @@ t_token	*create_word_token(char *input, int *index, char *unmatched_quote,
 	expanded_word = remove_quotes_and_expand(original_word, minishell);
 	if (!expanded_word)
 		return (free(original_word), minishell->exit_code = 1, NULL);
-	return (create_token_with_values(original_word, expanded_word, has_quotes));
+	return (create_token_with_values(original_word, expanded_word,
+		has_quotes));
 }
